mtf/metadata/requirements.txt
avocado-framework
pytest